requirements.txt
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
docutils==0.20.1
greenlet==3.0.3
idna==3.6
lockfile==0.12.2
luigi==3.5.0
lxml==5.1.0
mysql-connector-python==8.3.0
numpy==1.26.3
pandas==2.1.4
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
tenacity==8.2.3
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==2.1.0
